refactor: Exported delete button to utility file to be used on other pages
